feat(mod-searation): split connection.rs into multiple modules
